refactor: use @carbon/icons@10.8.0-rc.0
Remove devDeps: npm-run-all, prettier-plugin-svelte, svelte.
