+++ package.json
@@ -7,17 +7,14 @@
   "svelte": "lib/index.js",
   "main": "lib/index.js",
   "scripts": {
-    "build": "run-s build:*",
-    "build:task": "tsc",
-    "build:lib": "node dist",
-    "build:format-lib": "prettier --write 'lib/**/*.{js,svelte}'",
+    "build": "tsc && node dist",
     "test": "jest --coverage",
     "test:tdd": "jest --watch",
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
     "@carbon/icon-helpers": "10.4.0",
-    "@carbon/icons": "10.6.1",
+    "@carbon/icons": "10.8.0-rc.0",
     "@commitlint/cli": "^8.2.0",
     "@commitlint/config-conventional": "^8.2.0",
     "@types/fs-extra": "^8.0.1",
@@ -26,11 +23,8 @@
     "husky": "^3.1.0",
     "jest": "^24.9.0",
     "lint-staged": "^9.4.3",
-    "npm-run-all": "^4.1.5",
     "prettier": "^1.19.1",
-    "prettier-plugin-svelte": "^0.7.0",
     "pretty-quick": "^2.0.1",
-    "svelte": "^3.16.4",
     "ts-jest": "^24.2.0",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
@@ -49,6 +43,7 @@
     ]
   },
   "prettier": {
+    "printWidth": 100,
     "tabWidth": 2,
     "semi": true,
     "singleQuote": true
